Changed returntype of main to int
git-svn-id: https://svn.aros.org/svn/aros/trunk@6658 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/system/FixFonts.c
@@ -44,7 +44,7 @@
 char version[]="$VER: FixFonts 40.2 (16.9.99) (c) THOR";
 
 ///main
-long __asm __saveds main(void)
+int __asm __saveds main(void)
 {
 //struct ExecBase *SysBase;
 struct DosLibrary *DOSBase;
